Moved `append_to_sfile` up to base class
